+++ ErsatzTV/Pages/MultiSelectBase.cs
@@ -98,7 +98,8 @@ namespace ErsatzTV.Pages
             _selectedItems.OfType<TelevisionSeasonCardViewModel>().Map(s => s.TelevisionSeasonId).ToList(),
             _selectedItems.OfType<TelevisionEpisodeCardViewModel>().Map(e => e.EpisodeId).ToList(),
             _selectedItems.OfType<ArtistCardViewModel>().Map(a => a.ArtistId).ToList(),
-            _selectedItems.OfType<MusicVideoCardViewModel>().Map(mv => mv.MusicVideoId).ToList());
+            _selectedItems.OfType<MusicVideoCardViewModel>().Map(mv => mv.MusicVideoId).ToList(),
+            _selectedItems.OfType<OtherVideoCardViewModel>().Map(ov => ov.OtherVideoId).ToList());
 
         protected async Task AddItemsToCollection(
             List<int> movieIds,
@@ -107,10 +108,11 @@ namespace ErsatzTV.Pages
             List<int> episodeIds,
             List<int> artistIds,
             List<int> musicVideoIds,
+            List<int> otherVideoIds,
             string entityName = "selected items")
         {
             int count = movieIds.Count + showIds.Count + seasonIds.Count + episodeIds.Count + artistIds.Count +
-                        musicVideoIds.Count;
+                        musicVideoIds.Count + otherVideoIds.Count;
 
             var parameters = new DialogParameters
                 { { "EntityType", count.ToString() }, { "EntityName", entityName } };
@@ -127,7 +129,8 @@ namespace ErsatzTV.Pages
                     seasonIds,
                     episodeIds,
                     artistIds,
-                    musicVideoIds);
+                    musicVideoIds,
+                    otherVideoIds);
 
                 Either<BaseError, Unit> addResult = await Mediator.Send(request);
                 addResult.Match(